Initial commit: Gestion des commandes

+++ Models/Produit.cs
@@ -5,6 +5,6 @@ namespace GestionCommandes.Models
         public int Id { get; set; }
         public string Libelle { get; set; }
         public int QuantiteStock { get; set; }
-        public decimal PrixUnitaire { get; set; }
+        public double PrixUnitaire { get; set; }
     }
 }
